requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
google-generativeai==0.3.2
tenacity==8.2.3
PyPDF2==3.0.1
pytesseract==0.3.10
pillow==10.2.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2 